Revert "Stop saved Tiles from including unlicensed content"
This reverts commit dc9ae3d85b4ccc4739c1f628aace661d640f9089.

Using external resources isn’t the way to fix this problem. An external
resource must exist before the scene is loaded. The scene must be loaded
before the script starts, but the script is needed to generate the
external resource. Godot would crash before the script was able to
generate the files that Godot was complaining it couldn’t find.