Store Tile textures in an Array
Before this change, each face Texture had its own variable.
Unfortunately, that caused the Tile to run update_material() multiple
times when the scene was instantiated. I had added a workaround for that
problem, but that workaround caused problems with scenes that had been
instantiated but not added to the SceneTree.