Create Getting Started guide
It’s not done yet, but I thought that it would be better to commit what
I have so far. I could have waited until the guide is complete, but, in
my experience, trying to make something perfect before you release it
causes perfection to take longer. Plus, there’s a low chance that
someone will find this tool and try to use it before I finish the
Getting Started guide. It would be better to provide that person with
an incomplete guide than no guide at all.