Add contributing guide
I want the software that I create to be the best that it can possibly
be. The only way to achieve that goal is to get as many people as
possible to contribute. I suspect that giving people instructions for
how to contribute significantly increases the likely hood that people
will contribute.