Explicitly ask for contributions
I want my software to be the best that it can possibly be. In order for
my software to be the best that it can possibly be, I need other people
to contribute to it. I believe that explicitly asking for contributions
will significantly increase the likelihood that each user will
contribute.