Remove Util.TILE_IMAGE_FORMAT
List of problems with TILE_IMAGE_FORMAT:
1. It was an alias for another constant. It’s shorter to access that
constant than it was to access TILE_IMAGE_FORMAT.
2. The name is misleading. TILE_IMAGE_FORMAT was the image format used
for SingleColorTextures. SingleColorTextures can be applied to Tiles,
but they can also be used as floors or ceilings. Plus, you could
definitely apply a texture to a Tile has a different image format.
3. TILE_IMAGE_FORMAT was used in exactly one place, and that one place
wasn’t even in the file where TILE_IMAGE_FORMAT was defined.