Move make_dir_recursive_or_error() definition
make_dir_recursive_or_error() used to be in Util, but it was only ever
used in GraphicsExtractor. Originally, I think that I had put it in Util
in order to stay organized. The problem is that the code is a bit
tangled at the moment. There’s a significant number of classes that
can’t be preloaded or used via their class_name because it would create
a dependency cycle. I’m currently moving classes around to untangle the
code. Hopefully, untangling the code will allow me to (or at least make
it easier to) preload more classes.