Allow single threaded graphics extraction
Before I started working on this commit, I was working on another commit
that caused the GraphicsExtractor to fail. The failure didn’t happen in
the main thread, so it didn’t appear in the debugger. This change allows
me to run the code in the main thread so that I can see when it crashes.