Add BaseMap.api_version
At the moment, a Thing’s default type is 1 (player 1 start). That
default is not ideal. According to the UWMF spec, a thing’s type should
be a string not an integer [1]. Unfortunately, the latest stable version
of ECWolf (1.3.3) only supports thing types that are integers. The
latest development version of ECWolf supports using either integers or
strings, but using strings is deprecated and will stop working in the
future [2]. As a result, the Thing scene’s default type should
eventually be changed to a string.

Changing Thing’s default type (or the default value of any property) is
likely to cause a compatibility breakage. If you instantiate a child
scene, then each of that child scene’s properties will only be saved if
the property’s value is different from the default. In other words, maps
that have a Thing who’s type is 1 won’t save the fact that the type is
1. If we change the Thing scene’s default type to 666, then every Thing
who’s type is 1 will turn into a Thing who’s type is 666.

I don’t want updating Scene 2 UWMF to significantly change the results
of exporting a map. Changing minor details like the order in which tiles
are defined or the capitalization of properties is fine. What’s not fine
is changing kind of Actor is spawned or changing whether or not integers
are used.

The solution is to make default values dependent on the version of Scene
2 UWMF the map was designed for. This commit adds an api_version
property to BaseMap. If a scene changes any of its default properties in
the future, then it should bump BaseMap.LATEST_API_VERSION and only use
the new default value for maps with a high enough api_version (older
default values should be used with older api_versions).

[1]: <http://maniacsvault.net/ecwolf/wiki/UWMF#Required_Properties_5>
[2]: <http://maniacsvault.net/ecwolf/wiki/Version_compatibility#Upgrade_to_1.4_.28Development.29>