Rename Thing’s type property to type_number
This change is in preparation for when I start supporting using strings
for thing types. Using strings for thing types is what the UWMF spec
says to do [1] and will be required by a future version of ECWolf [2].
The only reason using integers for types was even supported in the first
place was because the latest stable version of ECWolf (1.3.3) requires
it.

[1]: <https://maniacsvault.net/ecwolf/wiki/UWMF#Things>
[2]: <https://maniacsvault.net/ecwolf/wiki/Version_compatibility#Upgrade_to_1.4_.28Development.29>